backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
pydantic-settings
python-dotenv
pytest
httpx
bcrypt
pyjwt[crypto]
python-multipart
email-validator
